#10 Create marketdef submodule, cache parsed definitions
